+++ R/parallel.R
@@ -1,6 +1,6 @@
-run_parallel_backend <- function(config) {
+run_drake_backend <- function(config) {
   get(
-    paste0("run_", config$parallelism),
+    paste0("backend_", config$parallelism),
     envir = getNamespace("drake")
   )(config)
 }
@@ -12,25 +12,25 @@ parallel_filter <- function(x, f, jobs = 1, ...) {
 }
 
 lightly_parallelize <- function(X, FUN, jobs = 1, ...) {
-  jobs <- safe_jobs_imports(jobs)
+  jobs <- safe_jobs(jobs)
   if (is.atomic(X)) {
     lightly_parallelize_atomic(X = X, FUN = FUN, jobs = jobs, ...)
   } else {
-    safe_mclapply(X = X, FUN = FUN, mc.cores = jobs, ...)
+    weak_mclapply(X = X, FUN = FUN, mc.cores = jobs, ...)
   }
 }
 
 lightly_parallelize_atomic <- function(X, FUN, jobs = 1, ...) {
-  jobs <- safe_jobs_imports(jobs)
+  jobs <- safe_jobs(jobs)
   keys <- unique(X)
   index <- match(X, keys)
-  values <- safe_mclapply(X = keys, FUN = FUN, mc.cores = jobs, ...)
+  values <- weak_mclapply(X = keys, FUN = FUN, mc.cores = jobs, ...)
   values[index]
 }
 
 # Avoid SIGCHLD handler when mc.cores is 1.
 # Could help avoid zeromq interrupted system call errors.
-safe_mclapply <- function(X, FUN, mc.cores, ...) {
+weak_mclapply <- function(X, FUN, mc.cores, ...) {
   if (mc.cores > 1) {
     parallel::mclapply(X = X, FUN = FUN, mc.cores = mc.cores, ...)
   } else {
@@ -38,31 +38,9 @@ safe_mclapply <- function(X, FUN, mc.cores, ...) {
   }
 }
 
-# x is parallelism or jobs
-imports_setting <- function(x) {
-  if (length(x) < 2) {
-    x
-  } else {
-    unname(x["imports"])
-  }
-}
-
-# x is parallelism or jobs
-targets_setting <- function(x) {
-  if (length(x) < 2) {
-    x
-  } else {
-    unname(x["targets"])
-  }
-}
-
 safe_jobs <- function(jobs) {
   stopifnot(length(jobs) == 1)
   ifelse(on_windows(), 1, jobs)
-}
-
-safe_jobs_imports <- function(jobs) {
-  ifelse(on_windows(), 1, imports_setting(jobs))
 }
 
 on_windows <- function() {
@@ -73,10 +51,80 @@ this_os <- function() {
   unname(tolower(Sys.info()["sysname"]))
 }
 
-parallelism_warnings <- function(config) {
-  warn_mclapply_windows(
-    parallelism = config$parallelism,
-    jobs = config$jobs,
-    os = this_os()
+#' @title Write a template file for deploying
+#'   work to a cluster / job scheduler.
+#' @description See the example files from
+#'  [drake_examples()] and [drake_example()]
+#'   for example usage.
+#' @export
+#' @seealso [drake_hpc_template_files()],
+#'   [drake_examples()], [drake_example()],
+#'   [shell_file()]
+#' @return `NULL` is returned,
+#'   but a batchtools template file is written.
+#' @param file Name of the template file, including the "tmpl" extension.
+#' @param to Character vector, where to write the file.
+#' @param overwrite Logical, whether to overwrite an existing file of the
+#'   same name.
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' load_mtcars_example() # Get the code with drake_example("mtcars").
+#' # List the available template files.
+#' drake_hpc_template_files()
+#' # Write a SLURM template file from the SLURM example.
+#' drake_hpc_template_file("slurm_batchtools.tmpl") # Writes slurm_batchtools.tmpl.
+#' # library(future.batchtools) # nolint
+#' # future::plan(batchtools_slurm, template = "slurm_batchtools.tmpl") # nolint
+#' # make(my_plan, parallelism = "future", jobs = 2) # nolint
+#' })
+#' }
+drake_hpc_template_file <- function(
+  file = drake::drake_hpc_template_files(),
+  to = getwd(),
+  overwrite = FALSE
+) {
+  file <- match.arg(file)
+  dir <- system.file(
+    "hpc_template_files",
+    package = "drake",
+    mustWork = TRUE
+  )
+  file.copy(from = file.path(dir, file), to = to,
+            overwrite = overwrite, recursive = TRUE)
+  invisible()
+}
+
+#' @title List the available example template files for deploying
+#'   work to a cluster / job scheduler.
+#' @description See the example files from
+#'  [drake_examples()] and [drake_example()]
+#'   for example usage.
+#' @export
+#' @seealso [drake_hpc_template_file()],
+#'   [drake_examples()], [drake_example()],
+#'   [shell_file()]
+#' @return A character vector of example template files that
+#'   you can write with [drake_hpc_template_file()].
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' load_mtcars_example() # Get the code with drake_example("mtcars").
+#' # List the available template files.
+#' drake_hpc_template_files()
+#' # Write a SLURM template file from the SLURM example.
+#' drake_hpc_template_file("slurm_batchtools.tmpl") # Writes slurm_batchtools.tmpl.
+#' # library(future.batchtools) # nolint
+#' # future::plan(batchtools_slurm, template = "slurm_batchtools.tmpl") # nolint
+#' # make(my_plan, parallelism = "future", jobs = 2) # nolint
+#' })
+#' }
+drake_hpc_template_files <- function() {
+  dir(
+    system.file(
+      "hpc_template_files",
+      package = "drake",
+      mustWork = TRUE
+    )
   )
 }