+++ tests/testthat/test-deprecate.R
@@ -84,6 +84,16 @@ test_with_dir("deprecate misc utilities", {
   expect_warning(config <- drake_config(my_plan, graph = 1, layout = 2))
   expect_warning(migrate_drake_project())
   expect_warning(default_verbose())
+  expect_warning(default_Makefile_args(2, 2))
+  expect_warning(default_Makefile_args(0, 0))
+  expect_warning(default_Makefile_command())
+  expect_warning(Makefile_recipe())
+  expect_warning(default_recipe_command())
+  expect_warning(r_recipe_wildcard())
+  expect_warning(parallelism_choices(TRUE))
+  expect_warning(parallelism_choices(FALSE))
+  expect_warning(shell_file())
+  expect_warning(default_parallelism())
 })
 
 test_with_dir("deprecated arguments", {
@@ -91,6 +101,7 @@ test_with_dir("deprecated arguments", {
   pl <- drake_plan(a = 1, b = a)
   con <- drake_config(plan = pl)
   expect_warning(drake_build(a, config = con, meta = list()))
+  expect_warning(make(drake_plan(x = 1), recipe_command = "123"))
 })
 
 test_with_dir("example template files (deprecated)", {